refactor: typify webhook and migrate from middleware pattern (#251)
* Move webhooks service into webhook module

* Rename webhooks.service to webhook.service

* Move webhook types into webhook module folder

* Move webhook.controller into modules file

* Move webhook utils to webhook modules folder

* Convert webhook controller to typescript

* Convert webhook controller to typescript

* Move webhook controller tests into util.spec and service.spec

* Refactor webhook service to return promises for testing purposes

* Convert to arrow function

* Remove declaration of submission model as compilation is done by the discriminator model itself

* Add types to webhook service spec

* Remove extra line and add jsdoc for response

* Add comment about moving away from middleware pattern

* Add comment that pushData should not be awaited upon

* Add JSDoc to webhook service functions

* Convert to await/async for better readability
