+++ src/types/index.ts
@@ -16,4 +16,3 @@ export * from './token'
 export * from './user'
 export * from './verification'
 export * from './admin_verification'
-export * from './webhook'
